Add procedure

+++ RazorWebApp/Pages/Employees/Edit.cshtml.cs
@@ -61,7 +61,10 @@ namespace RazorWebApp.Pages.Employees
                     {
                         string filePath = Path.Combine(_webHostEnvironment.WebRootPath, "images", Employee.PhotoPath);
 
-                        System.IO.File.Delete(filePath);
+                        if (Employee.PhotoPath != "noimage.png")
+                        {
+                            System.IO.File.Delete(filePath);
+                        }
                     }
 
                     Employee.PhotoPath = ProcessUploadedFile();
